Slideshare to PDF/requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
decorator==5.1.1
html5lib==1.1
idna==3.3
Pillow==9.0.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.8
validators==0.18.2
webencodings==0.5.1
